requirements.txt
langchain==0.2.16
langchain-community==0.2.16
pypdf==4.3.1
pinecone-client==5.0.1
langchain_pinecone==0.1.3
rank_bm25==0.2.2
openai== 1.44.0
streamlit==1.38.0
python-dotenv==1.0.1